Reorganize the scitokens-credmon condor config
Commented out and pulled CREDD_OAUTH_MODE and SEC_DEFAULT_ENCRYPTION
up into the "MANDATORY" block:
1. CREDD_OAUTH_MODE will cause the Schedd problems if enabled but if
the credmon is not running.
2. SEC_DEFAULT_ENCRYPTION should be part of a pool's security config,
but the option is still here if the admin just wants to uncomment it.

Added TRUST_CREDENTIAL_DIRECTORY, which is needed until we get a
better handle on the SEC_CREDENTIAL_DIRECTORY permissions.