Configs are not installed by default
Instead, configs will be installed programatically, like certbot.
This will allow the admin to modify paths/URLs *before* touching their
condor or apache configs as well as generate backups of existing
configs.

The credmon/utils/install.py code is pretty ugly right now and should
be improved so that any changes to example configs are reflected in the
actual configs (which are currently large strings in install.py).